+++ yarn-project/p2p/src/services/reqresp/batch-tx-requester/reqresp_batch.ts
@@ -3,26 +3,23 @@ import { createLogger } from '@aztec/foundation/log';
 import { Semaphore } from '@aztec/foundation/queue';
 import { sleep } from '@aztec/foundation/sleep';
 import type { BlockProposal } from '@aztec/stdlib/p2p';
-import { type Tx, type TxArray, TxHash } from '@aztec/stdlib/tx';
+import { type TxArray, TxHash } from '@aztec/stdlib/tx';
 
 import type { PeerId } from '@libp2p/interface';
 import { peerIdFromString } from '@libp2p/peer-id';
 
-import type { ConnectionSampler } from './connection-sampler/connection_sampler.js';
-import { type ReqRespInterface, ReqRespSubProtocol } from './interface.js';
-import { BlockTxsRequest, BlockTxsResponse } from './protocols/index.js';
-import { ReqRespStatus } from './status.js';
+import type { ConnectionSampler } from '.././connection-sampler/connection_sampler.js';
+import { type ReqRespInterface, ReqRespSubProtocol } from '.././interface.js';
+import { BlockTxsRequest, BlockTxsResponse } from '.././protocols/index.js';
+import { ReqRespStatus } from '.././status.js';
+import { MissingTxMetadata, MissingTxMetadataCollection, TX_BATCH_SIZE } from './missing_txs.js';
+import { PeerCollection } from './peer_colleciton.js';
 
-const TX_BATCH_SIZE = 8;
 const SMART_PEERS_TO_QUERY_IN_PARALLEL = 10;
 const DUMB_PEERS_TO_QUERY_IN_PARALLEL = 10;
 
 export class BatchTxRequester {
-  private readonly peers: PeerId[];
-  private readonly smartPeers = new Set<string>();
-  private readonly badPeers = new Map<string, number>();
-  private readonly inFlightPeers = new Set<string>();
-
+  private readonly peers: PeerCollection;
   private readonly txsMetadata;
 
   private readonly deadline;
@@ -41,12 +38,12 @@ export class BatchTxRequester {
     this.txsMetadata = new MissingTxMetadataCollection(
       missingTxs.map(txHash => [txHash.toString(), new MissingTxMetadata(txHash)]),
     );
-    this.peers = this.connectionSampler.getPeerListSortedByConnectionCountAsc();
+    this.peers = new PeerCollection(this.connectionSampler.getPeerListSortedByConnectionCountAsc());
     console.log(this.peers);
 
     //Pinned peer is queried separately and thus always considered "in-flight" by both "dumb" and "smart" requester
     if (this.pinnedPeer) {
-      this.inFlightPeers.add(pinnedPeer!.toString());
+      this.peers.markPeerInFlight(this.pinnedPeer);
     }
 
     this.deadline = Date.now() + this.timeoutMs;
@@ -91,7 +88,7 @@ export class BatchTxRequester {
 
   private async smartRequester() {
     const nextPeerIndex = this.makeRoundRobinIndexer();
-    const getPeers = () => Array.from(this.smartPeers.difference(this.getBadPeers().union(this.inFlightPeers)));
+    const getPeers = () => this.peers.getSmartPeersToQuery();
 
     const nextPeer = () => {
       const peers = getPeers();
@@ -110,19 +107,17 @@ export class BatchTxRequester {
       return { blockRequest: BlockTxsRequest.fromBlockProposalAndMissingTxs(this.blockProposal, txs), txs };
     };
 
-    const workers = Array.from({ length: Math.min(DUMB_PEERS_TO_QUERY_IN_PARALLEL, this.peers.length) }, () =>
-      this.workerLoop(nextPeer, makeRequest, 'smart'),
+    const workers = Array.from(
+      { length: Math.min(DUMB_PEERS_TO_QUERY_IN_PARALLEL, this.peers.getAllPeers().size) },
+      () => this.workerLoop(nextPeer, makeRequest, 'smart'),
     );
 
     await Promise.allSettled(workers);
   }
 
   private async dumbRequester() {
-    const peers = new Set(this.peers.map(peer => peer.toString()));
     const nextPeerIndex = this.makeRoundRobinIndexer();
     const nextBatchIndex = this.makeRoundRobinIndexer();
-    const getPeers = () =>
-      Array.from(peers.difference(this.smartPeers.union(this.getBadPeers()).union(this.inFlightPeers)));
 
     const txChunks = () =>
       //TODO: wrap around  for last batch
@@ -147,13 +142,14 @@ export class BatchTxRequester {
     };
 
     const nextPeer = () => {
-      const peers = getPeers();
+      const peers = this.peers.getDumbPeersToQuery();
       const idx = nextPeerIndex(() => peers.length);
       return idx === undefined ? undefined : peerIdFromString(peers[idx]);
     };
 
-    const workers = Array.from({ length: Math.min(DUMB_PEERS_TO_QUERY_IN_PARALLEL, this.peers.length) }, () =>
-      this.workerLoop(nextPeer, makeRequest, 'dumb'),
+    const workers = Array.from(
+      { length: Math.min(DUMB_PEERS_TO_QUERY_IN_PARALLEL, this.peers.getAllPeers().size) },
+      () => this.workerLoop(nextPeer, makeRequest, 'dumb'),
     );
     await Promise.allSettled(workers);
   }
@@ -207,7 +203,7 @@ export class BatchTxRequester {
 
   private async requestTxBatch(peerId: PeerId, request: BlockTxsRequest): Promise<BlockTxsResponse | undefined> {
     try {
-      this.inFlightPeers.add(peerId.toString());
+      this.peers.markPeerInFlight(peerId);
       const response = await this.reqresp.sendRequestToPeer(peerId, ReqRespSubProtocol.BLOCK_TXS, request.toBuffer());
       if (response.status !== ReqRespStatus.SUCCESS) {
         console.log(`Peer ${peerId.toString()} failed to respond with status: ${response.status}`);
@@ -226,12 +222,12 @@ export class BatchTxRequester {
       console.log(`Peer ${peerId.toString()}\n${err}`);
       await this.handleFailResponseFromPeer(peerId, ReqRespStatus.UNKNOWN);
     } finally {
-      this.inFlightPeers.delete(peerId.toString());
+      this.peers.unMarkPeerInFlight(peerId);
     }
   }
 
   private handleSuccessResponseFromPeer(peerId: PeerId, response: BlockTxsResponse) {
-    this.unMarkPeerAsBad(peerId);
+    this.peers.markPeerAsBad(peerId);
     this.logger.debug(`Received txs: ${response.txs.length} from peer ${peerId.toString()} `);
     this.handleReceivedTxs(peerId, response.txs);
 
@@ -247,9 +243,9 @@ export class BatchTxRequester {
     }
 
     this.markTxsPeerHas(peerId, response);
-    this.smartPeers.add(peerId.toString());
+    this.peers.markPeerSmart(peerId);
 
-    if (this.smartPeers.size <= SMART_PEERS_TO_QUERY_IN_PARALLEL) {
+    if (this.peers.getSmartPeers().size <= SMART_PEERS_TO_QUERY_IN_PARALLEL) {
       this.smartRequesterSemaphore.release();
     }
   }
@@ -261,7 +257,7 @@ export class BatchTxRequester {
     return blockIdsMatch && peerHasSomeTxsFromProposal;
   }
 
-  private peerHasSomeTxsWeAreMissing(peerId: PeerId, response: BlockTxsResponse): boolean {
+  private peerHasSomeTxsWeAreMissing(_peerId: PeerId, response: BlockTxsResponse): boolean {
     const txsPeerHas = new Set(this.extractHashesPeerHasFromResponse(response).map(h => h.toString()));
     return this.txsMetadata.getMissingTxHashes().intersection(txsPeerHas).size > 0;
   }
@@ -283,31 +279,13 @@ export class BatchTxRequester {
   // banning the peers?
   private async handleFailResponseFromPeer(peerId: PeerId, responseStatus: ReqRespStatus) {
     if (responseStatus === ReqRespStatus.FAILURE || responseStatus === ReqRespStatus.UNKNOWN) {
-      this.markPeerAsBad(peerId);
+      this.peers.markPeerAsBad(peerId);
     }
 
     //TODO: handle this properly
     if (responseStatus === ReqRespStatus.RATE_LIMIT_EXCEEDED) {
       await sleep(1000);
     }
-  }
-
-  private markPeerAsBad(peerId: PeerId) {
-    this.badPeers.set(peerId.toString(), (this.badPeers.get(peerId.toString()) ?? 0) + 1);
-  }
-
-  private unMarkPeerAsBad(peerId: PeerId) {
-    this.badPeers.delete(peerId.toString());
-  }
-
-  private getBadPeers(): Set<string> {
-    const BADE_PEER_THRESHOLD = 3;
-    return new Set(
-      this.badPeers
-        .entries()
-        .filter(([_k, v]) => v > BADE_PEER_THRESHOLD)
-        .map(([k]) => k),
-    );
   }
 
   private extractHashesPeerHasFromResponse(response: BlockTxsResponse): Array<TxHash> {
@@ -343,192 +321,5 @@ export class BatchTxRequester {
   //TODO: abort signal here?
   private shouldStop() {
     return this.txsMetadata.size === 0 || this.fetchedAllTxs() || Date.now() > this.deadline;
-  }
-}
-
-class MissingTxMetadata {
-  constructor(
-    public readonly txHash: TxHash,
-    public fetched = false,
-    public requestedCount = 0,
-    public inFlightCount = 0,
-    public tx: Tx | undefined = undefined,
-    public readonly peers = new Set<string>(),
-  ) {}
-
-  public markAsRequested() {
-    this.requestedCount++;
-  }
-
-  public markInFlight() {
-    this.inFlightCount++;
-  }
-
-  public markNotInFlight() {
-    this.inFlightCount = Math.max(--this.inFlightCount, 0);
-  }
-
-  public isInFlight(): boolean {
-    return this.inFlightCount > 0;
-  }
-
-  public markAsFetched(peerId: PeerId, tx: Tx) {
-    this.fetched = true;
-    this.tx = tx;
-
-    this.peers.add(peerId.toString());
-  }
-
-  public toString() {
-    return this.txHash.toString();
-  }
-}
-
-/*
- * Single source or truth for transactions we are fetching
- * This could be better optimized but given expected count of missing txs (N < 100)
- * At the moment there is no need for it. And benefit is that we have everything in single store*/
-class MissingTxMetadataCollection extends Map<string, MissingTxMetadata> {
-  public getSortedByRequestedCountAsc(txs: string[]): MissingTxMetadata[] {
-    return Array.from(this.values().filter(txMeta => txs.includes(txMeta.txHash.toString()))).sort(
-      (a, b) => a.requestedCount - b.requestedCount,
-    );
-  }
-
-  public getSortedByRequestedCountThenByInFlightCountAsc(txs: string[]): MissingTxMetadata[] {
-    const filtered = Array.from(this.values()).filter(txMeta => txs.includes(txMeta.txHash.toString()));
-
-    const [notInFlight, inFlight] = filtered.reduce<[MissingTxMetadata[], MissingTxMetadata[]]>(
-      (buckets, tx) => {
-        tx.isInFlight() ? buckets[1].push(tx) : buckets[0].push(tx);
-        return buckets;
-      },
-      [[], []],
-    );
-
-    notInFlight.sort((a, b) => a.requestedCount - b.requestedCount);
-    inFlight.sort((a, b) => a.inFlightCount - b.inFlightCount);
-
-    return [...notInFlight, ...inFlight];
-  }
-
-  public isFetched(txHash: TxHash): boolean {
-    //If something went' wrong and we don't have txMeta for this hash
-    // We should not request it, so here we "pretend" that it was fetched
-    return this.get(txHash.toString())?.fetched ?? true;
-  }
-
-  public getFetchedTxHashes(): Set<string> {
-    return new Set(
-      this.values()
-        .filter(t => t.fetched)
-        .map(t => t.txHash.toString()),
-    );
-  }
-
-  public getMissingTxHashes(): Set<string> {
-    return new Set(
-      this.values()
-        .filter(t => !t.fetched)
-        .map(t => t.txHash.toString()),
-    );
-  }
-
-  public getInFlightTxHashes(): Set<string> {
-    return new Set(
-      this.values()
-        .filter(t => t.isInFlight())
-        .map(t => t.txHash.toString()),
-    );
-  }
-
-  public getFetchedTxs(): Tx[] {
-    return Array.from(
-      this.values()
-        .map(t => t.tx)
-        .filter(t => !!t),
-    );
-  }
-
-  public getTxsPeerHas(peer: PeerId): Set<string> {
-    const peerIdStr = peer.toString();
-    const txsPeerHas = new Set<string>();
-
-    this.values().forEach(txMeta => {
-      if (txMeta.peers.has(peerIdStr)) {
-        txsPeerHas.add(txMeta.txHash.toString());
-      }
-    });
-
-    return txsPeerHas;
-  }
-
-  public getTxsToRequestFromThePeer(peer: PeerId): TxHash[] {
-    const txsPeerHas = this.getTxsPeerHas(peer);
-    const fetchedTxs = this.getFetchedTxHashes();
-
-    const txsToRequest = txsPeerHas.difference(fetchedTxs);
-    if (txsToRequest.size >= TX_BATCH_SIZE) {
-      return this.getSortedByRequestedCountThenByInFlightCountAsc(Array.from(txsToRequest)).map(t => t.txHash);
-    }
-
-    // Otherwise fill the txs to request till TX_BATCH_SIZE with random txs we are missing
-    // Who knows, maybe we get lucky and peer received these txs in the meantime
-
-    const countToFill = TX_BATCH_SIZE - txsToRequest.size;
-    const txsToFill = this.getSortedByRequestedCountThenByInFlightCountAsc(
-      Array.from(this.getMissingTxHashes().difference(txsToRequest)),
-    )
-      .slice(0, countToFill)
-      .map(t => t.txHash);
-
-    return [...Array.from(txsToRequest).map(t => TxHash.fromString(t)), ...txsToFill];
-  }
-
-  public markRequested(txHash: TxHash) {
-    this.get(txHash.toString())?.markAsRequested();
-  }
-
-  /*
-   * This should be called only when requesting tx from smart peer
-   * Because the smart peer should return this tx, whereas
-   * "dumb" peer might return it, or might not - we don't know*/
-  public markInFlightBySmartPeer(txHash: TxHash) {
-    this.get(txHash.toString())?.markInFlight();
-  }
-
-  /*
-   * This should be called only when requesting tx from smart peer
-   * Because the smart peer should return this tx, whereas
-   * "dumb" peer might return it, or might not - we don't know*/
-  public markNotInFlightBySmartPeer(txHash: TxHash) {
-    this.get(txHash.toString())?.markNotInFlight();
-  }
-
-  public markFetched(peerId: PeerId, tx: Tx) {
-    const txHashStr = tx.txHash.toString();
-    const txMeta = this.get(txHashStr);
-    if (!txMeta) {
-      //TODO: what to do about peer which sent txs we didn't request?
-      // 1. don't request from it in the scope of this batch request
-      // 2. ban it immediately?
-      // 3. track it and ban it?
-      //
-      return;
-    }
-
-    txMeta.markAsFetched(peerId, tx);
-  }
-
-  public markPeerHas(peerId: PeerId, txHash: TxHash[]) {
-    const peerIdStr = peerId.toString();
-    txHash
-      .map(t => t.toString())
-      .forEach(txh => {
-        const txMeta = this.get(txh);
-        if (txMeta) {
-          txMeta.peers.add(peerIdStr);
-        }
-      });
   }
 }
